clean & update few things

+++ app/Livewire/Settings/ThirdPartyAccount.php
@@ -2,8 +2,6 @@
 
 namespace App\Livewire\Settings;
 
-use App\Enums\ProviderName;
-use App\Http\Controllers\Auth\GoogleController;
 use App\Http\Controllers\Auth\IdentityProviderController;
 use App\Models\IdentityProvider;
 use Illuminate\Support\Collection;
@@ -40,13 +38,12 @@ class ThirdPartyAccount extends Component
 
     public function unlinkProvider()
     {
+        if (!$this->selectedProvider) return;
+
         $this->validate([
             'password'  => ['required', 'string', 'current_password'],
         ]);
 
-        // if ($this->selectedProvider->provider_name == ProviderName::GOOGLE) {
-        //     (new GoogleController)->unlink();
-        // }
         (new IdentityProviderController)->unlink(provider: $this->selectedProvider->provider_name);
 
         return to_route('third-party-account.edit')->with('sso-succeded', 'Successfully unlinked your '.$this->selectedProvider->provider_name->label().' account.');
